refactor: refactoring solving code
