+++ src/pages/solving/components/SolvingHeader.tsx
