+++ src/pages/solving/components/SolvingQuizTitle.tsx
@@ -7,6 +7,10 @@ interface SolvingQuizTitleProps {
 	type?: "molu";
 }
 
+//
+//
+//
+
 const SolvingQuizTitle = ({ title, type = "molu" }: SolvingQuizTitleProps) => {
 	const lines = title.split("\n|\r");
