+++ src/pages/solving/pages/solving-quiz-content/SolvingQuizContentMultipleAnswer.tsx
@@ -1,0 +1,16 @@
+import SolvingAnswerLayout from "../../layouts/SolvingAnswerLayout";
+
+//
+//
+//
+
+const SolvingQuizContentMultipleAnswer = () => {
+	const answers = Array.from({ length: 4 }).map((_, index) => ({
+		number: index + 1,
+		value: `${index + 1}번 문제의 답변.`,
+	}));
+
+	return <SolvingAnswerLayout answers={answers} />;
+};
+
+export default SolvingQuizContentMultipleAnswer;
