+++ src/pages/solving/pages/solving-quiz-content/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SolvingQuizContent";
