+++ src/pages/solving/pages/solving-top-bar/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SolvingTopBar";
